% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postregression_functions.R
\name{cjoint_plot}
\alias{cjoint_plot}
\title{Plot a FactorHet Object}
\usage{
cjoint_plot(object, baseline = NA)
}
\arguments{
\item{object}{An object from FactorHet or FactorHet_mbo.}

\item{baseline}{Default is "NA" and shows all coefficients. Providing a named
list of baseline levels to difference each main effect from. See help for
\code{marginal_AME}.}
}
\value{
\item{plot}{The ggplot2 visualization of the marginal effects.}
\item{data}{The data used to create the ggplot2 visualization.}
}
\description{
Plots the regression coefficients beta_k from a fitted FactorHet object for
main effects only. Use \code{marginal_AME} to calculate average marginal effects
on a predicted probability scale.
}
\examples{
data(immigration)
# Induce "fake" randomization restriction
immigration$joint_id <- paste(immigration$CaseID, immigration$contest_no)
remove_profiles <- subset(immigration, Plans == 'No plans' & Ed == 'GradDeg')
immigration <- subset(immigration, !(joint_id \%in\% remove_profiles$joint_id))
# Fit with one cluster and limited regularization for example only
# Ignore conjoint structure for simplicity
fit_MBO <- FactorHet(Chosen_Immigrant ~ Plans + Ed + Country,
  design = immigration, lambda = 1e-4,
  K = 1, group = ~ CaseID, task = ~ contest_no, choice_order = ~ choice_id)
# Plot the raw coefficients
cjoint_plot(fit_MBO)

}
\seealso{
\link{marginal_AME}
}
