% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_penalty.R
\name{create_penalty}
\alias{create_penalty}
\title{Create penalty matrix (list of F)}
\usage{
create_penalty(
  factor_levels,
  make_interactions,
  weight_dlist,
  ordered_factors,
  treatment_probs = NULL
)
}
\arguments{
\item{factor_levels}{A list with the levels of each factor}

\item{treatment_probs}{A list with the levels of each factor and the probability of treatment assignment}
}
\value{
A list of penalty matrices (F) used for sparsification
}
\description{
Create penalty matrix (list of F)
}
\keyword{internal}
