% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postregression_functions.R,
%   R/visualization_functions.R
\name{marginal_AME}
\alias{marginal_AME}
\alias{marginal_ACE}
\alias{marginal_AMIE}
\alias{moderator_AME}
\title{Deprecated Functions}
\usage{
marginal_AME(...)

marginal_ACE(...)

marginal_AMIE(...)

moderator_AME(...)
}
\value{
Returns objects corresponding to those described in
  \code{\link{AME}}, \code{\link{ACE}}, \code{\link{AMIE}}, and
  \code{\link{margeff_moderators}}, respectively.
}
\description{
The names of these functions have changed, e.g., \code{AME} to
\code{AME}, for clarity. Please adjust your code to use the current names.
}
\seealso{
\link{AME} \link{margeff_moderators}
}
\keyword{internal}
