% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_functions.R
\encoding{UTF-8}
\name{FactorHet_control}
\alias{FactorHet_control}
\title{Control for FactorHet Estimation}
\usage{
FactorHet_control(
  iterations = 1000,
  maxit_pi = 25,
  optim_phi_controls = NULL,
  prior_var_phi = 4,
  prior_var_beta = Inf,
  gamma = 1,
  repeat_beta = 1,
  adaptive_weight = "B&R",
  init_method = "short_EM",
  return_data = FALSE,
  log_method = "log_ginv",
  tolerance.parameters = 1e-04,
  tolerance.logposterior = 1e-05,
  rare_threshold = 5,
  rare_verbose = 1,
  beta_method = "cg",
  beta_cg_it = 25,
  lambda_scale = "N",
  weight_dlist = FALSE,
  do_SQUAREM = TRUE,
  backtrack_SQUAREM = 3,
  df_method = "EM",
  forced_randomize = FALSE,
  single_intercept = NULL,
  tau_method = "nullspace",
  tau_stabilization = 5,
  tau_truncate = 1e+06,
  debug = FALSE,
  calc_df = TRUE,
  calc_se = TRUE,
  quiet_tictoc = TRUE,
  override_BR = FALSE
)
}
\arguments{
\item{iterations}{Number of iterations for the algorithm.}

\item{maxit_pi}{Number of steps of L-BFGS-B to perform. Default of 25.}

\item{optim_phi_controls}{Options to \link{optim} to numerically optimize
moderator parameters.}

\item{prior_var_phi}{Variance of multivariate normal prior on moderator
coefficients.}

\item{prior_var_beta}{Variance of normal prior on each treatment effect
coefficient. Set to \code{Inf} by default when using sparse estimation.}

\item{gamma}{Determines whether sparsity-inducing prior be "spread" across
clusters in proportion to the average prior probability of membership. See
the Städler et al. (2010) for discussion.}

\item{repeat_beta}{Default of "1"; higher numbers repeat the "E-M" step for
updating \code{beta} multiple times before performing optimization of
moderator parameters.}

\item{adaptive_weight}{Should weights be given to the penalties? By default
it imposes Bondell and Reich (2009) weights (\code{"B&R"}). Please see
Goplerud et al. (2022) for discussion. If a matrix is provided (e.g. from a
prior run of \code{FactorHet}, this can be used to set up an "adaptive
overlapping group LASSO". \code{"none"} imposes no weights. To use a matrix
and \emph{not} use Bondell and Reich weights, set \code{override_BR =
TRUE}.}

\item{init_method}{Method for initializing model. Can accept a string of
\code{"kmeans"} (k-means clustering on the moderators), \code{"mclust"}
(\code{"mclust"} on the moderators), \code{"random_pi"} (random
probabilities of cluster membership for each person),
\code{"random_member"} (random hard assignment), \code{"random_beta"}
(random coefficients). More commonly, this can be set with a named list
with cluster membership probabilities. This should consist of a named list
with a single element \code{"group_E.prob"} that is a data.frame which
contains probabilities for each group/unit with the column names
\code{"group"} and then \code{"cluster_[0-9]+"} depending on \code{K}. A
full set of parameters can be provided, although this is only used
internally for debugging.
be set using the options in \link{FactorHet_mbo_control} as most users will
calibrate the regularization using \code{FactorHet_mbo}.}

\item{return_data}{Return the formatted data; default of \code{FALSE}.}

\item{log_method}{Method for dealing with latent overlapping groups. "log_"
implements latent overlapping groups (see Yan and Bien 2017; see also the
supporting information of Goplerud et al. 2022). The projection matrix can
be either the generalized inverse extending Post and Bondell (2013)
(default: \code{"log_ginv"}), a random matrix (\code{"log_random"}), or
zero (\code{"log_0"}). \code{"standard"} does not implement overlapping
groups.}

\item{tolerance.parameters}{Convergence criterion when no parameter changes
by more than this amount.}

\item{tolerance.logposterior}{Convergence criterion when the log-posterior
changes by less than this amount.}

\item{rare_threshold}{Remove interactions with fewer than this many
observations.}

\item{rare_verbose}{Print information about factors removed.}

\item{beta_method}{Method for updating beta, e.g. "cpp" or "base R".}

\item{beta_cg_it}{If using conjugate gradient, how many steps?}

\item{lambda_scale}{Function for internally rescaling lambda to be a function
of N. Options are \code{"N"} (default; \code{lambda * N}), \code{"unity"}
(i.e. no rescaling), or \code{"root_N"} (\code{lambda * sqrt(N)}).}

\item{weight_dlist}{Weight additional penalties following Hastie and Lim
(2015)? Default of \code{FALSE}}

\item{do_SQUAREM}{Perform SQUAREM to accelerate convergence. Strongly
recommended to keep as \code{TRUE}.}

\item{backtrack_SQUAREM}{The number of backtracking steps to perform for
SQUAREM.}

\item{df_method}{Method for calculating degrees of freedom. Default of
\code{"EM"} follows Goplerud et al. (2022) and calculates the degrees of
freedom using the Polya-Gamma weights. \code{"IRLS"} uses p_{ik} (1 -
p_{ik}) as weights, where p_{ik} = Pr(y_i = 1 | z_i = k).
\code{"free_param"} counts the number of parameters after fusion and
accounting for the sum-to-zero constraints. Use \code{"all"} to estimate
all methods and compare.}

\item{forced_randomize}{For each task, randomize which is "left" and "right".
Default of \code{FALSE}.}

\item{single_intercept}{Should a single intercept be used across clusters?
Default of \code{NULL} uses a single intercept if forced-choice and a
varying intercept if factorial.}

\item{tau_method}{Method for dealing with binding restrictions, i.e.
numerically infinite E[1/tau^2] The two options are \code{"nullspace"}
(i.e. perform inference assuming this restriction binds) or \code{"clip"}
(set to a large value.) \code{"nullspace"} is preferred.}

\item{tau_stabilization}{Perform this many steps of the algorithm with
\code{"clip"} (see \code{tau_method}) to avoid initial instability.}

\item{tau_truncate}{Value to either clip tau (i.e. set maximum E[1/tau^2] or
set in nullspace depending on \code{tau_method}).}

\item{debug}{Verify log-posterior at each step and show size of nullspace.}

\item{calc_df}{Calculate degrees of freedom of final model. Default of
\code{TRUE}.}

\item{calc_se}{Calculate standard errors of final model. Default of
\code{TRUE}.}

\item{quiet_tictoc}{Do not print timing information from tictoc.}

\item{override_BR}{Ignore Bondell and Reich style-weights. Instead, use
\code{sqrt(L) * (L + 1)} where \code{L} is the number of factor levels.}
}
\description{
Set control arguments for estimation FactorHet. Arguments around the
initialization of the model (important when \code{K > 1}) can be set via
\link{FactorHet_init} and arguments for the model-based optimization tuning
of \code{lambda} can be found in \link{FactorHet_mbo_control}. The parameters
can be divided into ones governing the model priors, model estimation, and
miscellaneous settings. All arguments have default values.
}
\examples{
str(FactorHet_control())

}
\references{
Bondell, Howard D., and Brian J. Reich. 2009. "Simultaneous Factor Selection
and Collapsing Levels in ANOVA." Biometrics 65(1): 169-177.

Goplerud, Max, Kosuke Imai, and Nicole E. Pashley. 2022. "Estimating
Heterogeneous Causal Effects of High-Dimensional Treatments: Application to
Conjoint Analysis." arxiv preprint: \url{https://arxiv.org/abs/2201.01357}

Post, Justin B., and Howard D. Bondell. 2013. "Factor Selection and
Structural Identification in the Interaction ANOVA Model." \emph{Biometrics}
69(1):70-79.

Lim, Michael, and Trevor Hastie. 2015. "Learning Interactions via Hierarchical
Group-Lasso Regularization." \emph{Journal of Computational and Graphical
Statistics} 24(3):627-654.

Städler, Nicolas, Peter Bühlmann, and Sara Van De Geer. 2010.
"l1-penalization for Mixture Regression Models." \emph{Test} 19(2):209-256.

Yan, Xiaohan and Jacob Bien. 2017. "Hierarchical Sparse Modeling: A Choice of
Two Group Lasso Formulations." \emph{Statistical Science} 32(4):531–560.
}
