% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_functions.R
\encoding{UTF-8}
\name{FactorHet_init}
\alias{FactorHet_init}
\title{Arguments for Initializing FactorHet}
\usage{
FactorHet_init(
  short_EM = FALSE,
  short_EM_it = 40,
  short_EM_init = "random_member",
  short_EM_pi = 50,
  force_rep = FALSE,
  verbose = FALSE,
  short_EM_beta_method = "cg",
  short_EM_cg_it = 10,
  nrep = 5,
  debug_repeat = F,
  plot_repeat = F,
  return_all = F
)
}
\arguments{
\item{short_EM}{Logical. \code{TRUE} indicates a "short EM" should be
followed. That is, run multiple short runs of EM with random
initializations and then proceed with the best for full initialization.
See, e.g., Biernacki et al. (2003) for more discussion.}

\item{short_EM_it}{Number of iterations to use for the short EM.}

\item{short_EM_init}{Initialization procedure for "short EM". Must be some
non-deterministic procedure that is valid in \code{FactorHet_control}.}

\item{short_EM_pi}{Number of iterations to use to update moderators.}

\item{force_rep}{Force repeated restarts when K = 1; only used for debugging.}

\item{verbose}{Print progress of estimation.}

\item{short_EM_beta_method}{Method to update betas.}

\item{short_EM_cg_it}{Number of iterations for conjugate gradient.}

\item{nrep}{How many random initializations/short EMs should be used?}

\item{debug_repeat}{Debug the repeated option.}

\item{plot_repeat}{Plot the trajectory of the log-posterior for each run}

\item{return_all}{Return all models.}
}
\description{
A set of arguments to affect the initalization of FactorHet. Use
\link{FactorHet_control} to set arguments around estimation.
\link{FactorHet_mbo} ignores many of these arguments as it uses a single
fixed initialization for each proposed regularization value. All arguments
have default values.
}
\examples{
str(FactorHet_init())

}
\references{
Biernacki, Christophe, Gilles Celeux, and Gérard Govaert. "Choosing Starting
Values for the EM algorithm for Getting the Highest Likelihood in
Multivariate Gaussian Mixture Models." 2003. \emph{Computational Statistics &
Data Analysis}. 41(3-4):561-575.
}
