% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{moderator_AME}
\alias{moderator_AME}
\title{Visualize the effects of moderators}
\usage{
moderator_AME(
  object,
  newdata = NULL,
  vcov = TRUE,
  se.method = NULL,
  quant_continuous = c(0.25, 0.75)
)
}
\arguments{
\item{object}{Object fit using \code{FactorHet} or \code{FactorHet_mbo}.}

\item{newdata}{Default \code{NULL} marginalizes over the estimation data.
Provide a data.frame with the relevant moderators to calculate the average
effects of moderators changing on that data.}

\item{vcov}{Default of \code{TRUE} shows standard errors.}

\item{se.method}{Default of \code{NULL} uses estimated standard errors. See
\code{FactorHet.vcov} for more information.}

\item{quant_continuous}{For continuous moderator, two quantiles to show
effect difference between.}
}
\description{
Visualize the effects of moderators on cluster membership. Provides a graph
of the posterior (predictive) distribution as well as the marginal effect of
changing each of the moderators. For each function, assigning the output to
an object returns the raw underlying data used to create the figure.
}
