% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_functions.R
\name{prepare_regression_data}
\alias{prepare_regression_data}
\title{Prepare Data}
\usage{
prepare_regression_data(
  formula,
  design,
  moderator = NULL,
  group = NULL,
  task = NULL,
  choice_order = NULL,
  weights = NULL,
  single_intercept = NULL,
  forced_randomize = NULL,
  rare_threshold = NULL,
  weight_dlist = FALSE,
  rare_verbose = NULL
)
}
\description{
Package data to use in FactorHet. Called internally by `FactorHet_mbo`;
rarely called directly by the user.
}
\keyword{internal}
