% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{rank_via_null}
\alias{rank_via_null}
\title{Rank of Matrix}
\usage{
rank_via_null(X, outer = FALSE)
}
\arguments{
\item{X}{Sparse Matrix with size N by p.}

\item{outer}{Calculate rank of X^T X. Often much faster if N >> p.}
}
\description{
Calculate rank of (sparse) matrix by a QR decomposition.
}
\keyword{internal}
