% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{FactorHet-class}
\alias{FactorHet-class}
\alias{plot.FactorHet}
\alias{formula.FactorHet}
\alias{print.FactorHet}
\alias{summary.FactorHet}
\alias{coef.FactorHet}
\alias{logLik.FactorHet}
\alias{BIC.FactorHet}
\alias{AIC.FactorHet}
\alias{print.FactorHet_vis}
\alias{visualize_MBO}
\alias{posterior_FactorHet}
\title{Generic Methods for FactorHet models}
\usage{
\method{plot}{FactorHet}(x, y = NULL, ...)

\method{formula}{FactorHet}(x, ...)

\method{print}{FactorHet}(x, fusion.tolerance = 1e-04, ...)

\method{summary}{FactorHet}(object, show_interactions = FALSE, digits = 3, ...)

\method{coef}{FactorHet}(object, coef_type = "beta", ...)

\method{logLik}{FactorHet}(object, type = "loglik", ...)

\method{BIC}{FactorHet}(object, ...)

\method{AIC}{FactorHet}(object, ...)

\method{print}{FactorHet_vis}(x, ...)

visualize_MBO(object)

posterior_FactorHet(object)
}
\arguments{
\item{x}{Model from FactorHet}

\item{y}{Not used; required to maintain compatability.}

\item{...}{Optional arguments; only used by \code{plot.FactorHet} to pass
arguments to \code{cjoint_plot}.}

\item{fusion.tolerance}{Threshold at which to declare levels fused}

\item{object}{Object fit using \code{FactorHet} or \code{FactorHet_mbo}.}

\item{show_interactions}{Used by \code{summary.FactorHet}; should the
interaction terms be shown? Default \code{FALSE}. See "Details" for more
discussion.}

\item{digits}{Number of digits to include}

\item{coef_type}{Type of coefficient (beta for treatment effects; phi for moderators)}

\item{type}{For "logLik", should the log-likelihood (\code{"loglik"}),
log-posterior (\code{"log_posterior"}), or sequence of log-posterior values
at each iteration (\code{"log_posterior_seq"}) be returned?}
}
\description{
Brief descriptions of generic methods (e.g. print, summary) for FactorHet as
well as a way to visualize the progress of the model-based optimization.
}
\details{
The following methods with the arguments given above exist. All
  methods work on models with using \code{FactorHet} and
  \code{FactorHet_mbo}. 
  \itemize{
  \item{plot: }{A shorthand for \link{cjoint_plot} on a fitted object.}
  \item{formula: }{Get the underlying formula for the treatment effects and
  moderators. This also returns the values used for \code{group},
  \code{task: }, and \code{choice_order} if provided.}
  \item{print: }{Two print methods. For \code{FactorHet}, it summarizes the
  model and fusion of the factor levels. \code{fusion.tolerance} sets the
  threshold at which levels are reported as fused. For outputs of
  \link{marginal_AME} (and similar), this plots the corresponding plot. See
  that documentation for more details.}
  \item{summary: }{Summarizes the main effects by cluster with standard errors.
  It is typically more common to visualize this with \link{cjoint_plot} (and
  the accompanying data.frame) or \link{marginal_AME}.
  \code{show_interactions = TRUE} shows the interactions in addition to the
  main effects.}
  \item{coef: }{Returns the coefficient matrix on the original scale (i.e. with
  the sum-to-zero constraints). \code{code_type = "phi"} returns the
  moderator coefficients instead of the treatment effect coefficients.}
  \item{AIC and BIC: }{Reports the AIC or BIC. If multiple degrees of freedom
  options specified, returns a matrix. Otherwise, it returns a number.}
  \item{logLik: }{Returns the log-likelihood, log-posterior or sequence of
  log-posterior values at each iteration of the algorithm. See argument
  \code{type} above for details.}
  \item{visualize_MBO: }{For a model fit with \code{FactorHet_mbo}, show
  information about the MBO, i.e. proposed values and objectives.}
  \item{posterior_FactorHet: }{For a model with \code{K > 1}, visualize the
  posterior for each observation and the posterior predictive implied by the
  moderators. The output is a list of class \code{FactorHet_vis} that
  contains the posterior and posterior-predictive probabilities in the
  argument \code{"data"}. This data.frame has four columns: "group" for the
  individual i's id, "post.predict" for the posterior predictive probability
  (the probability of cluster membership if no data were observed), i.e.
  \eqn{\pi_{ik}}, and the posterior probability (the probability of cluster
  membership from the EM algorithm after observing the data), i.e.
  \eqn{E[z_ik]}.}
 }
}
