% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorHet.R, R/mbo_functions.R
\name{FactorHet}
\alias{FactorHet}
\alias{FactorHet_mbo}
\title{Estimating Heterogeneous Effects in Factorial and Conjoint Experiments}
\usage{
FactorHet(
  formula,
  design,
  K,
  lambda,
  moderator = NULL,
  group = NULL,
  task = NULL,
  choice_order = NULL,
  weights = NULL,
  control = FactorHet_control(),
  initialize = FactorHet_init(),
  verbose = TRUE
)

FactorHet_mbo(
  formula,
  design,
  K,
  moderator = NULL,
  weights = NULL,
  group = NULL,
  task = NULL,
  choice_order = NULL,
  control = FactorHet_control(),
  initialize = FactorHet_init(),
  mbo_control = FactorHet_mbo_control()
)
}
\arguments{
\item{formula}{Formula specifying model. The syntax is y ~ X1 + X2 where y is
the outcome and X1 and X2 are factors. Interactions can be specified using
* syntax. All main factors must be explicitly included.}

\item{design}{A data.frame containing the experimental data.}

\item{K}{The number of clusters to estimate.}

\item{lambda}{The regularization strength; it is not common to set this
manually versus using FactorHet_mbo to optimize this term.}

\item{moderator}{A formula of variables (moderators) that affect the prior
probability of cluster membership. Ignored when K = 1.}

\item{group}{A formula of a single variable, e.g. \code{~ person_id}, that is
used when there are repeated observations per individual.}

\item{task}{A formula of a single variable that indicates the task number
performed by each individual. This is not used when \code{group} is unspecified.}

\item{choice_order}{A formula of a single variable that indicates which
profile is on the "left" or "right" in a conjoint experiment.}

\item{weights}{A vector of weights for each observation (e.g. survey
weights). If group is specified, they must be constant inside of each value
of group.}

\item{control}{An object from \link{FactorHet_control} that sets various model
estimation options.}

\item{initialize}{An object from \link{FactorHet_init} that determines how the
model is initialized.}

\item{verbose}{A logical term that prints intermediate information about
model fitting.}

\item{mbo_control}{A list of control parameters for MBO; see
\link{FactorHet_mbo_control} for more information.}
}
\value{
Returns an object of class \code{FactorHet}. Typical use will involve
  examining the patterns of estimated treatment effects. \code{cjoint_plot} shows
  the raw (logistic) coefficients. 
  
  Marginal effects (e.g. average marginal
  effect) can be computed using \code{marginal_AME}, \code{marginal_ACE}, or
  \code{marginal_AMIE}. 
  
  The effects of moderators can be examined using
  \code{moderator_AME} or \code{posterior_by_moderators}. Please see the corresponding
  documentation for details.
  
  The returned object is a list containing the following elements:
  \itemize{
  \item{parameters: }{Estimated model parameters. Usually extract these via
  \code{coef}.}
  \item{K: }{The number of clusters}
  \item{posterior: }{Posterior cluster probability for each observation; list
  of two data.frames one with posterior (\code{"posterior"}) and one
  (\code{"posterior_predictive"}) based on the probabilities \eqn{\pi_{ik}} implied
  by the moderators.}
  \item{information_criterion: }{Information on the information criterion estiamted.}
  \item{internal_parameters: }{Internal parameters; for debugging or use by other functions}
  \item{vcov: }{Named list containing the estimated variance-covariance
  matrix, usually extracted with \code{vcov}. Other objects for internal
  use.}
  \item{lp_shortEM: }{If "short EM" is applied, the log-posterior at the end
  of each short run.}
  \item{MBO:}{If \code{FactorHet_mbo} is used, information about the
  model-based optimization is stored here. Use \link{visualize_MBO} to see a
  graphical summary.}
  }
}
\description{
Fit a model to estimate heterogeneous effects in factorial or conjoint
experiments using a "mixture of experts" (i.e. a finite mixture of
regularized regressions with covariates affecting cluster assignment).
Effects are regularized using an overlapping group LASSO.
}
\details{
All models are estimated using an AECM algorithm described in Goplerud et al.
(2022).

\bold{Estimation:} Calibration of the amount of regularization
  should be done using \code{FactorHet_mbo}. This uses some number of
  attempts to calibrate the amount of regularization by minimizing some
  criterion (e.g. the BIC), and then fits a final model using the optimal
  estimated regularization.

  Options for the model based optimization (mbo) can be set using
  \code{FactorHet_mbo_control()}. Options for model estimation can be set
  using \code{FactorHet_control()}.
  
  \bold{Ridge Regression:} While more experimental, ridge regression can be
  estimated by setting \code{lambda = 0} (in \code{FactorHet}) and then
  setting \code{prior_var_beta} in \code{FactorHet_control} or by using
  \code{FactorHet_mbo} and setting \code{mbo_type = "ridge"}.
  
  \bold{Moderators:} Moderators can be provided via the \code{moderator}
  argument. These are important when \code{K > 1} for ensuring the stability
  of the model. Repeated observations per individual can be specified by
  \code{group} and/or \code{task} if relevant for a force-choiced conjoint.
}
\examples{
# Use a small subset of the immigration data from Hainmueller and Hopkins
data(immigration)
# Fit with two clusters and tune regularization via MBO
# Only do one iteration (iters = 1) for speed
fit_MBO <- FactorHet_mbo(
  formula = Chosen_Immigrant ~ Country + Ed + Gender + Plans,
  design = immigration, group = ~ CaseID,
  task =  ~ contest_no, choice_order = ~ choice_id,
  # Only do one guess after initialization for speed
  mbo_control = FactorHet_mbo_control(iters = 1),
  K = 2)
# Plot the raw coefficients
cjoint_plot(fit_MBO)
# Check how MBO fared at calibrating regularization
visualize_MBO(fit_MBO)
# Visualize posterior distribution of cluster membership
posterior_FactorHet(fit_MBO)
# Get AMEs
marginal_AME(fit_MBO)
}
